extension-tools: Add type definitions and update documentation
Add `extension-tools/types.d.ts` with TypeScript type definitions for
extension tool authors. This enables IDE autocompletion and type
checking via JSDoc comments in plain JavaScript files.

Update `ExtensionToolLoader.ts` to import types from the shared
`types.d.ts` instead of defining them locally, ensuring consistency
between the loader and the public API.

Also update sample tools (`get_google_calendar.js`, `get_tasks.js`)
with JSDoc type annotations demonstrating the recommended pattern.

Written by Claude
