+++ extension-tools/get_google_calendar.js
@@ -232,8 +232,10 @@ async function fetchEvents(requestUrl, startDate, endDate, log, warn) {
   return filtered;
 }
 
+/** @param {import('./types').ExtensionToolContext} ctx */
 module.exports = function (ctx) {
-  return {
+  /** @type {import('./types').ExtensionTool} */
+  const tool = {
     definition: {
       name: 'get_google_calendar',
       description:
@@ -275,4 +277,5 @@ module.exports = function (ctx) {
       return formatEventsForContext(events);
     },
   };
+  return tool;
 };
